[UPDATE]: emergency mobile fix

+++ flutter_application/lib/ServicePage/action_reaction_page.dart
@@ -200,7 +200,7 @@ class CreateActionReactionPage extends StatelessWidget {
           child: Column(
             children: <Widget>[
               Padding(
-                padding: EdgeInsets.only(left: sidebarWidth),
+                padding: EdgeInsets.only(left: sidebarWidth.toDouble()),
                 child: CreateActionReactionCards(),
               ),
             ],
